Moved @ValidUUID to general verification package.
